laravel authentication and login

+++ resources/views/student.blade.php
@@ -50,9 +50,17 @@
                    <td>{{ $data->nis }}</td> 
                    <td>{{ $data->class->name }}</td>
                    <td>
-                        <a href="student/{{ $data->id }}">Detail</a>
-                        <a href="student-edit/{{ $data->id }}">Edit</a>
-                        <a href="student-delete/{{ $data->id }}">Delete</a>
+                        @if (Auth::user()->role_id != 1 && Auth::user()->role_id !=2)
+                            -
+                        @else
+                            <a href="student/{{ $data->id }}">Detail</a>
+                            <a href="student-edit/{{ $data->id }}">Edit</a>
+                        @endif
+
+                        @if (Auth::user()->role_id == 1)
+                            <a href="student-delete/{{ $data->id }}">Delete</a>
+                            
+                        @endif
                    </td>
                    {{-- <td>{{ $data->class['name'] }}</td> 
                    <td>
